Add abstraction for db
