--- src/Inspector/ActiveRecord/Common.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Debug\Api\Inspector\ActiveRecord;
-
-use Yiisoft\ActiveRecord\ActiveRecord;
-
-class Common extends ActiveRecord
-{
-}
